made a super class because I found a function that was common among three of our classes (help, gamepreferences, and playermenu control classes all had a function inside to display the main menu)
however, a better way to do this might be to simply call the display function from the mainmenuview instead of even implementing a seperate function to display this each time.  I think we have an extra step here.  Either way, it allowed me something to do and show for our individual programming assignment this week